+++ Makefile
@@ -10,8 +10,8 @@ GIT_COMMIT := $(shell git rev-parse HEAD)
 all: run
 
 run:
-	go build -ldflags="-X '$(PROJECT)/version.Version=$(VERSION)' \
-	-X '$(PROJECT)/version.Commit=$(GIT_COMMIT)'" -o $(APP_NAME) && ./$(APP_NAME)
+	go build -ldflags="-X '$(PROJECT)/internal/version.Version=$(VERSION)' \
+	-X '$(PROJECT)/internal/version.Commit=$(GIT_COMMIT)'" -o $(APP_NAME) ./cmd/static-srv/main.go && ./$(APP_NAME)
 
 build_container:
 	docker build --build-arg GIT_COMMIT=$(GIT_COMMIT) --build-arg VERSION=$(VERSION)  --build-arg=PROJECT=$(PROJECT) \
