+++ README.md
@@ -1,2 +1,5 @@
 # backend2
 домашки по курсу "backend 2"
+
+1. Создайте Makefile и добавьте в него все необходимые команды для компиляции, сборки и запуска контейнера. Выделите хелсчеки в отдельные пробы.
+2. Выберите любой сервис, написанный вами ранее. Проверьте, соответствует ли он практикам, изученным в предыдущем и текущем уроках. Разверните его в удалённый кластер. Подготовьте манифесты для всех стратегий развёртывания, изученных в этом уроке: Rolling Update, Fixed Deployment, Blue-Green Deployment, Canary Deployment.
