+++ k8s_manifest.yaml
@@ -1,0 +1,79 @@
+apiVersion: apps/v1
+kind: Deployment
+metadata:
+  name: static-srv
+  labels:
+    app: static-srv
+spec:
+  replicas: 3
+  strategy:
+    type: RollingUpdate
+    rollingUpdate:
+      maxUnavailable: 50%
+      maxSurge: 1
+  template:
+    metadata:
+      labels:
+        app: static-srv
+    spec:
+      volumes:
+        - name: src
+          emptyDir: { }
+      initContainers:
+        - name: prepare-statics
+          image: alpine/git
+          command:
+          - git
+          - clone
+          - https://github.com/mdn/beginner-html-site-scripted
+          - /static
+          volumeMounts:
+          - mountPath: /static
+            name: src
+      containers:
+        - name: static-srv
+          image: docker.io/seggga/static-srv:1.0.2
+          imagePullPolicy: Always
+          volumeMounts:
+            - mountPath: /static
+              name: src
+          env:
+           - name: PORT
+             value: "8080"
+           - name: STATICS_PATH
+             value: "/static"
+          ports:
+            - containerPort: 8080
+          livenessProbe:
+            httpGet:
+              path: /__heartbeat__
+              port: 8080
+          readinessProbe:
+            httpGet:
+              path: /__readiness__
+              port: 8080
+          resources:
+            limits:
+              cpu: 2m
+              memory: 10Mi
+            requests:
+              cpu: 2m
+              memory: 10Mi
+  selector:
+    matchLabels:
+      app: static-srv
+---
+apiVersion: v1
+kind: Service
+metadata:
+  name: static-srv
+  labels:
+    app: static-srv
+spec:
+  ports:
+    - port: 8080
+      targetPort: 8080
+      protocol: TCP
+      name: http
+  selector:
+    app: static-srv
